Internal merge

+++ src/com/bsb/hike/filetransfer/FTApkManager.java
@@ -37,7 +37,7 @@ import java.util.Calendar;
 public class FTApkManager
 {
 
-    private  static Context context;
+    private  static Context context = HikeMessengerApp.getInstance().getApplicationContext();
 
     public static void makeRequest(final double apkSizeMultiplier)
     {
@@ -66,7 +66,7 @@ public class FTApkManager
                     apkSize = apkJson.getLong(HikeConstants.AutoApkDownload.DOWNLOAD_APK_SIZE);
                     String[] parts = apkDownloadUrl.split("/");
                     String apkName = parts[3];
-                    apkName += ("_" + System.currentTimeMillis());
+                    apkName += ("_" + System.currentTimeMillis()) + ".apk";
 
                     long freeSdCardSpace = PhoneSpecUtils.getSDCardMem().get("free");
                     boolean spaceAvailable = freeSdCardSpace > (apkSizeMultiplier * apkSize);
@@ -98,7 +98,9 @@ public class FTApkManager
                             Logger.d("AUTOAPK", "on receive of new apk, the version name was found to be higher, hence cancelling and downloading");
                             FileTransferManager.getInstance(context).cancelTask(-100, hfOld.getFile(), false, hfOld.getFileSize());
                             Logger.d("AUTOAPK", "also deleting the already downloaded file if it exists");
-                            hfOld.getFile().delete();
+                            if (hfOld.getFile() != null && hfOld.getFile().exists()) {
+                                hfOld.getFile().delete();
+                            }
 
                             Logger.d("AUTOAPK", " download will start after network check, saving vargs");
                             HikeFile hf = new HikeFile(apkName, HikeFile.HikeFileType.toString(HikeFile.HikeFileType.APK), apkVersion, apkSize, apkDownloadUrl);
@@ -118,6 +120,13 @@ public class FTApkManager
                             if (networkType >= networkTypeAvailable && networkTypeAvailable > 0) {
                                 Logger.d("AUTOAPK", "Starting download now, correct network detected");
                                 FileTransferManager.getInstance(context).downloadApk(hf.getFile(), hf.getFileKey(), hf.getHikeFileType());
+
+                                JSONObject metadata = new JSONObject();
+                                metadata.put(HikeConstants.EVENT_TYPE, HikeConstants.MqttMessageTypes.AUTO_APK);
+                                metadata.put(HikeConstants.EVENT_KEY, AnalyticsConstants.AutoApkEvents.INITIATING_DOWNLOAD);
+                                metadata.put(AnalyticsConstants.AutoApkEvents.NETWORK_VALIDITY, networkTypeAvailable);
+
+                                HAManager.getInstance().record(AnalyticsConstants.NON_UI_EVENT, AnalyticsConstants.ANALYTICS_EVENT, HAManager.EventPriority.HIGH, metadata);
                             }
                         } else {
                             Logger.d("AUTOAPK", "not downloading the new APK, the version check failed, might continue with older apk");
@@ -176,7 +185,10 @@ public class FTApkManager
                 if (!jo.toString().equals("{}")) {
                     Logger.d("AUTOAPK", "hike file : " + jo.toString());
                     HikeFile hikefile = new HikeFile(jo, false);
-                    hikefile.getFile().delete();
+                    if(hikefile.getFile() != null && hikefile.getFile().exists())
+                    {
+                        hikefile.getFile().delete();
+                    }
                 }
             } catch (JSONException je)
             {
@@ -212,8 +224,8 @@ public class FTApkManager
             if(!jo.toString().equals("{}")) {
                 Logger.d("AUTOAPK", "hike file : " + jo.toString());
                 HikeFile hikefile = new HikeFile(jo, false);
-                File apk = hikefile.getFileFromExactFilePath();
-                File apkExact = hikefile.getFile();
+                File apkExact = hikefile.getFileFromExactFilePath();
+                File apk = hikefile.getFile();
                 if (apk != null) {
                     Logger.d("AUTOAPK", "hike file paths  : " + apk.getAbsolutePath());
                 }
@@ -222,18 +234,18 @@ public class FTApkManager
                 }
 
                 long apkSizeReceived = HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_SIZE, 0l);
-                boolean validSize = hikefile.getFile().length() == apkSizeReceived && hikefile.getFile().length() > 0;
-                Logger.d("AUTOAPK", hikefile.getFile().length() + " is the file size, saved size is " + apkSizeReceived);
+                boolean validSize =  (apk!=null && apk.exists() ) ? apk.length() == apkSizeReceived && apk.length() > 0 : false;
+                Logger.d("AUTOAPK", "Is the apk size valid ? : "+validSize  + ", saved size is " + apkSizeReceived);
                 boolean updateNeeded = Utils.isUpdateRequired(HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_VERSION, ""), context)
                         && !TextUtils.isEmpty(HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_VERSION, ""));
                 Logger.d("AUTOAPK", HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_VERSION, "") + " is the apk version new, old is is " + Utils.getAppVersionName());
 
-                if (hikefile.getFile().exists() && validSize && updateNeeded) {
+                if (apk != null && apk.exists() && validSize && updateNeeded) {
                     Logger.d("AUTOAPK", "hike APK downloaded exists");
                     if (HikeFile.HikeFileType.APK == hikefile.getHikeFileType()) {
                         Logger.d("AUTOAPK", "hike showing install prompt");
                         Intent intent = new Intent(Intent.ACTION_VIEW);
-                        intent.setDataAndType(Uri.fromFile(hikefile.getFile()), "application/vnd.android.package-archive");
+                        intent.setDataAndType(Uri.fromFile(apk), "application/vnd.android.package-archive");
                         intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
                         context.startActivity(intent);
                         if(isNonPlayAppAllowed)
@@ -263,11 +275,12 @@ public class FTApkManager
             JSONObject jo = new JSONObject(HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_JSON, "{}"));
             HikeFile hikeApkFile = new HikeFile(jo, false);
             long apkSizeReceived = HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_SIZE, 0l);
-            boolean validSize = hikeApkFile.getFile().length() == apkSizeReceived && hikeApkFile.getFile().length() > 0;
+            File hFile = hikeApkFile.getFile();
+            boolean validSize = (hFile != null && hFile.exists()) ?  hFile.length() == apkSizeReceived && hFile.length() > 0 : false;
             boolean updateNeeded = Utils.isUpdateRequired(HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_VERSION, ""), context)
                     && !TextUtils.isEmpty(HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_VERSION, ""));
-
-            if(hikeApkFile.getFile().equals(mFile))
+            boolean fileExists = hFile!=null && hFile.exists() && hFile.equals(mFile);
+            if(fileExists)
             {
                 if(validSize && updateNeeded)
                 {
@@ -281,7 +294,7 @@ public class FTApkManager
                 else
                 {
                     Logger.d("AUTOAPK", "deleting the file now, since file is not valid for installation");
-                    hikeApkFile.getFile().delete();
+                    hFile.delete();
                     HikeSharedPreferenceUtil pref = HikeSharedPreferenceUtil.getInstance();
                     pref.removeData(HikeConstants.AutoApkDownload.NEW_APK_JSON);
                     pref.removeData(HikeConstants.AutoApkDownload.NEW_APK_TIP_JSON);
@@ -296,6 +309,20 @@ public class FTApkManager
                     pref.removeData(HikeConstants.AutoApkDownload.UPDATE_FROM_DOWNLOADED_APK);
                 }
             }
+            else
+            {
+                Logger.d("AUTOAPK", "size validity : " + validSize + ", update Needed : " + updateNeeded);
+            }
+
+
+            JSONObject metadata = new JSONObject();
+            metadata.put(HikeConstants.EVENT_TYPE, HikeConstants.MqttMessageTypes.AUTO_APK);
+            metadata.put(HikeConstants.EVENT_KEY, AnalyticsConstants.AutoApkEvents.DOWNLOAD_COMPLETION);
+            metadata.put(AnalyticsConstants.AutoApkEvents.SIZE_VALIDITY, validSize);
+            metadata.put(AnalyticsConstants.AutoApkEvents.UPDATE_VALIDITY, updateNeeded);
+            metadata.put(AnalyticsConstants.AutoApkEvents.FILE_VALIDITY, fileExists);
+            HAManager.getInstance().record(AnalyticsConstants.NON_UI_EVENT, AnalyticsConstants.ANALYTICS_EVENT, HAManager.EventPriority.HIGH, metadata);
+
         }
         catch (JSONException je)
         {
@@ -341,7 +368,9 @@ public class FTApkManager
 
 
                     Logger.d("AUTOAPK", "delete apk and vars after updation has been perfromed");
-                    hikefile.getFile().delete();
+                    if(hikefile.getFile() != null && hikefile.getFile().exists()) {
+                        hikefile.getFile().delete();
+                    }
 
                 }
             } catch (JSONException je)
@@ -378,7 +407,7 @@ public class FTApkManager
             if(!jo.toString().equals("{}")) {
                 HikeFile hf = new HikeFile(jo, false);
 
-                String stamp = hf.getFileName().substring(hf.getFileName().length() - 13, hf.getFileName().length());
+                String stamp = hf.getFileName().substring(hf.getFileName().length() - 17, hf.getFileName().length() - 4);
                 Logger.d("AUTOAPK", stamp);
                 Long timeStamp = Long.parseLong(stamp);
                 Logger.d("AUTOAPK", timeStamp + "");
@@ -394,7 +423,9 @@ public class FTApkManager
                     FileTransferManager.getInstance(context).cancelTask(-100, hf.getFile(), false, hf.getFileSize());
 
                     Logger.d("AUTOAPK", "delete apk and vars on connection retry");
-                    hf.getFile().delete();
+                    if(hf.getFile()!=null && hf.getFile().exists()) {
+                        hf.getFile().delete();
+                    }
                     mprefs.removeData(HikeConstants.AutoApkDownload.NEW_APK_JSON);
                     mprefs.removeData(HikeConstants.AutoApkDownload.UPDATE_FROM_DOWNLOADED_APK);
                     mprefs.removeData(HikeConstants.AutoApkDownload.NEW_APK_TIP_JSON);
@@ -415,11 +446,26 @@ public class FTApkManager
 
 
                     long apkSizeReceived = HikeSharedPreferenceUtil.getInstance().getData(HikeConstants.AutoApkDownload.NEW_APK_SIZE, 0l);
-                    boolean downloadPending = hf.getFile().length() < apkSizeReceived;
+                    boolean downloadPending;
+                    if(hf.getFile()!=null && hf.getFile().exists())
+                    {
+                        downloadPending =  hf.getFile().length() < apkSizeReceived;
+                    }
+                    else
+                    {
+                        downloadPending = true;
+                    }
                     // TODO a better logic still pending for checking if download is required again, other wise it might cause multiple downloads
                     //TODO possible solution is : on download complete one should switch off the network change receiver
                     if (networkType >= networkTypeAvailable && networkTypeAvailable > 0 && downloadPending) {
                         Logger.d("AUTOAPK", "Starting download now, correct network detected");
+
+                        JSONObject metadata = new JSONObject();
+                        metadata.put(HikeConstants.EVENT_TYPE, HikeConstants.MqttMessageTypes.AUTO_APK);
+                        metadata.put(HikeConstants.EVENT_KEY, AnalyticsConstants.AutoApkEvents.RESUMING_DOWNLOAD);
+                        metadata.put(AnalyticsConstants.AutoApkEvents.NETWORK_VALIDITY, networkTypeAvailable);
+                        HAManager.getInstance().record(AnalyticsConstants.NON_UI_EVENT, AnalyticsConstants.ANALYTICS_EVENT, HAManager.EventPriority.HIGH, metadata);
+
                         FileTransferManager.getInstance(context).downloadApk(hf.getFile(), hf.getFileKey(), hf.getHikeFileType());
                     }
                 }}
@@ -427,7 +473,11 @@ public class FTApkManager
         }
         catch (JSONException je)
         {
-            Logger.d("AUTOAPK","json exception");
+            Logger.d("AUTOAPK","json exception" + je.getMessage());
+        }
+        catch (NumberFormatException nfe)
+        {
+            Logger.d("AUTOAPK","number format exception on parsing long");
         }
     }
